+++ packages/program-test-suite/src/faker.ts
@@ -1,0 +1,3 @@
+import {TriggerType} from '@saasquatch/program-boilerplate';
+
+export function getTriggerJson(type: TriggerType) {}
